ai-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skincare-analyzer-ai-agent"
version = "1.0.0"
description = "AI agent for skincare ingredient extraction and rating"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "pillow>=10.3.0",
    "pytesseract>=0.3.10",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.103.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "httpx>=0.27.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
